pyproject.toml
[project]
name = "wallofx"
version = "0.1.0"
description = "A Telegram bot that transforms tweets into beautiful printable images"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-telegram-bot>=21.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.1",
]

[project.scripts]
wallofx = "wallofx.src.bot:WallOfXBot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
